+++ sklearn/decomposition/pca.py
@@ -18,8 +18,9 @@ from scipy.special import gammaln
 from ..base import BaseEstimator, TransformerMixin
 from ..utils import array2d, check_random_state, as_float_array
 from ..utils import atleast2d_or_csr
-from ..utils.extmath import fast_logdet, safe_sparse_dot, randomized_svd, \
-                            fast_dot
+from ..utils import deprecated
+from ..utils.extmath import (fast_logdet, safe_sparse_dot, randomized_svd,
+                             fast_dot)
 
 
 def _assess_dimension_(spectrum, rank, n_samples, n_features):
@@ -153,10 +154,23 @@ class PCA(BaseEstimator, TransformerMixin):
         to 'mle' or a number between 0 and 1 to select using explained
         variance.
 
+    `noise_variance_` : float
+        The estimated noise covariance following the Probabilistic PCA model
+        from Tipping and Bishop 1999. See "Pattern Recognition and
+        Machine Learning" by C. Bishop, 12.2.1 p. 574 or
+        http://www.miketipping.com/papers/met-mppca.pdf. It is required to
+        computed the estimated data covariance and score samples.
+
     Notes
     -----
     For n_components='mle', this class uses the method of `Thomas P. Minka:
     Automatic Choice of Dimensionality for PCA. NIPS 2000: 598-604`
+
+    Implements the probabilistic PCA model from:
+    M. Tipping and C. Bishop, Probabilistic Principal Component Analysis,
+    Journal of the Royal Statistical Society, Series B, 61, Part 3, pp. 611-622
+    via the score and score_samples methods.
+    See http://www.miketipping.com/papers/met-mppca.pdf
 
     Due to implementation subtleties of the Singular Value Decomposition (SVD),
     which is used in this implementation, running fit twice on the same matrix
@@ -253,14 +267,14 @@ class PCA(BaseEstimator, TransformerMixin):
         self.mean_ = np.mean(X, axis=0)
         X -= self.mean_
         U, S, V = linalg.svd(X, full_matrices=False)
-        self.explained_variance_ = (S ** 2) / n_samples
-        self.explained_variance_ratio_ = (self.explained_variance_ /
-                                          self.explained_variance_.sum())
+        explained_variance_ = (S ** 2) / n_samples
+        explained_variance_ratio_ = (explained_variance_ /
+                                     explained_variance_.sum())
 
         if self.whiten:
-            self.components_ = V / S[:, np.newaxis] * sqrt(n_samples)
+            components_ = V / (S[:, np.newaxis] / sqrt(n_samples))
         else:
-            self.components_ = V
+            components_ = V
 
         n_components = self.n_components
         if n_components is None:
@@ -269,23 +283,86 @@ class PCA(BaseEstimator, TransformerMixin):
             if n_samples < n_features:
                 raise ValueError("n_components='mle' is only supported "
                                  "if n_samples >= n_features")
-            n_components = _infer_dimension_(self.explained_variance_,
+
+            n_components = _infer_dimension_(explained_variance_,
                                              n_samples, n_features)
 
         if 0 < n_components < 1.0:
             # number of components for which the cumulated explained variance
             # percentage is superior to the desired threshold
-            ratio_cumsum = self.explained_variance_ratio_.cumsum()
+            ratio_cumsum = explained_variance_ratio_.cumsum()
             n_components = np.sum(ratio_cumsum < n_components) + 1
 
-        self.components_ = self.components_[:n_components, :]
-        self.explained_variance_ = \
-            self.explained_variance_[:n_components]
-        self.explained_variance_ratio_ = \
-            self.explained_variance_ratio_[:n_components]
+        # Compute noise covariance using Probabilistic PCA model
+        # The sigma2 maximum likelihood (cf. eq. 12.46)
+        if n_components < n_features:
+            self.noise_variance_ = explained_variance_[n_components:].mean()
+        else:
+            self.noise_variance_ = 0.
 
+        # store n_samples to revert whitening when getting covariance
+        self.n_samples_ = n_samples
+
+        self.components_ = components_[:n_components]
+        self.explained_variance_ = explained_variance_[:n_components]
+        explained_variance_ratio_ = explained_variance_ratio_[:n_components]
+        self.explained_variance_ratio_ = explained_variance_ratio_
         self.n_components_ = n_components
+
         return (U, S, V)
+
+    def get_covariance(self):
+        """Compute data covariance with the generative model.
+
+        ``cov = components_.T * S**2 * components_ + sigma2 * eye(n_features)``
+        where  S**2 contains the explained variances.
+
+        Returns
+        -------
+        cov : array, shape=(n_features, n_features)
+            Estimated covariance of data.
+        """
+        components_ = self.components_
+        exp_var = self.explained_variance_
+        if self.whiten:
+            components_ = components_ * np.sqrt(exp_var[:, np.newaxis])
+        exp_var_diff = np.maximum(exp_var - self.noise_variance_, 0.)
+        cov = np.dot(components_.T * exp_var_diff, components_)
+        cov.flat[::len(cov) + 1] += self.noise_variance_  # modify diag inplace
+        return cov
+
+    def get_precision(self):
+        """Compute data precision matrix with the generative model.
+
+        Equals the inverse of the covariance but computed with
+        the matrix inversion lemma for efficiency.
+
+        Returns
+        -------
+        precision : array, shape=(n_features, n_features)
+            Estimated precision of data.
+        """
+        n_features = self.components_.shape[1]
+
+        # handle corner cases first
+        if self.n_components_ == 0:
+            return np.eye(n_features) / self.noise_variance_
+        if self.n_components_ == n_features:
+            return linalg.inv(self.get_covariance())
+
+        # Get precision using matrix inversion lemma
+        components_ = self.components_
+        exp_var = self.explained_variance_
+        if self.whiten:
+            components_ = components_ * np.sqrt(exp_var[:, np.newaxis])
+        exp_var_diff = np.maximum(exp_var - self.noise_variance_, 0.)
+        precision = np.dot(components_, components_.T) / self.noise_variance_
+        precision.flat[::len(precision) + 1] += 1. / exp_var_diff
+        precision = np.dot(components_.T,
+                           np.dot(linalg.inv(precision), components_))
+        precision /= -(self.noise_variance_ ** 2)
+        precision.flat[::len(precision) + 1] += 1. / self.noise_variance_
+        return precision
 
     def transform(self, X):
         """Apply the dimensionality reduction on X.
@@ -328,7 +405,57 @@ class PCA(BaseEstimator, TransformerMixin):
         """
         return fast_dot(X, self.components_) + self.mean_
 
+    def score_samples(self, X):
+        """Return the log-likelihood of each sample
 
+        See. "Pattern Recognition and Machine Learning"
+        by C. Bishop, 12.2.1 p. 574
+        or http://www.miketipping.com/papers/met-mppca.pdf
+
+        Parameters
+        ----------
+        X: array, shape(n_samples, n_features)
+            The data.
+
+        Returns
+        -------
+        ll: array, shape (n_samples,)
+            Log-likelihood of each sample under the current model
+        """
+        Xr = X - self.mean_
+        n_features = X.shape[1]
+        log_like = np.zeros(X.shape[0])
+        precision = self.get_precision()
+        log_like = -.5 * (Xr * (np.dot(Xr, precision))).sum(axis=1)
+        log_like -= .5 * (n_features * log(2. * np.pi)
+                          - fast_logdet(precision))
+        return log_like
+
+    def score(self, X, y=None):
+        """Return the average log-likelihood of all samples
+
+        See. "Pattern Recognition and Machine Learning"
+        by C. Bishop, 12.2.1 p. 574
+        or http://www.miketipping.com/papers/met-mppca.pdf
+
+        Parameters
+        ----------
+        X: array, shape(n_samples, n_features)
+            The data.
+
+        Returns
+        -------
+        ll: float
+            Average log-likelihood of the samples under the current model
+        """
+        return np.mean(self.score_samples(X))
+
+
+@deprecated("ProbabilisticPCA will be removed in 0.16. WARNING: the covariance"
+            " estimation was previously incorrect, your output might be different "
+            " than under the previous versions. Use PCA that implements score"
+            " and score_samples. To work with homoscedastic=False, you should use"
+            " FactorAnalysis.")
 class ProbabilisticPCA(PCA):
     """Additional layer on top of PCA that adds a probabilistic evaluation"""
     __doc__ += PCA.__doc__
@@ -345,25 +472,28 @@ class ProbabilisticPCA(PCA):
             If True, average variance across remaining dimensions
         """
         PCA.fit(self, X)
-        n_samples, n_features = X.shape
-        self._dim = n_features
-        Xr = X - self.mean_
-        Xr -= np.dot(np.dot(Xr, self.components_.T), self.components_)
 
+        n_samples, n_features = X.shape
         n_components = self.n_components
         if n_components is None:
             n_components = n_features
 
+        explained_variance = self.explained_variance_.copy()
+        if homoscedastic:
+            explained_variance -= self.noise_variance_
+
         # Make the low rank part of the estimated covariance
         self.covariance_ = np.dot(self.components_[:n_components].T *
-                                  self.explained_variance_[:n_components],
+                                  explained_variance,
                                   self.components_[:n_components])
 
         if n_features == n_components:
             delta = 0.
         elif homoscedastic:
-            delta = (Xr ** 2).sum() / (n_samples * n_features)
+            delta = self.noise_variance_
         else:
+            Xr = X - self.mean_
+            Xr -= np.dot(np.dot(Xr, self.components_.T), self.components_)
             delta = (Xr ** 2).mean(axis=0) / (n_features - n_components)
 
         # Add delta to the diagonal without extra allocation